cleanup and refactor in order to get ready for future work
